+++ packages/datastore/src/authModeStrategies/multiAuthStrategy.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { Auth } from '@aws-amplify/auth';
+import { InternalAuth } from '@aws-amplify/auth/internals';
 import { GRAPHQL_AUTH_MODE } from '@aws-amplify/api-graphql';
 import {
 	AuthModeStrategy,
@@ -9,6 +9,7 @@ import {
 	ModelAttributeAuthAllow,
 	AmplifyContext,
 } from '../types';
+import { Category, DataStoreAction } from '@aws-amplify/core';
 
 function getProviderFromRule(
 	rule: ModelAttributeAuthProperty
@@ -139,11 +140,14 @@ export const multiAuthStrategy: (
 	amplifyContext: AmplifyContext
 ) => AuthModeStrategy =
 	(amplifyContext: AmplifyContext) =>
-	async ({ schema, modelName }) => {
-		amplifyContext.Auth = amplifyContext.Auth || Auth;
+	async ({ schema, modelName, customUserAgentDetails }) => {
+		amplifyContext.InternalAuth = amplifyContext.InternalAuth || InternalAuth;
 		let currentUser;
 		try {
-			currentUser = await amplifyContext.Auth.currentAuthenticatedUser();
+			currentUser = await amplifyContext.InternalAuth.currentAuthenticatedUser(
+				undefined,
+				customUserAgentDetails
+			);
 		} catch (e) {
 			// No current user
 		}