+++ packages/api-rest/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 3.5.2 (2023-08-17)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
+## 3.5.1 (2023-08-10)
+
+**Note:** Version bump only for package @aws-amplify/api-rest
+
 # [3.5.0](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-rest@3.4.0...@aws-amplify/api-rest@3.5.0) (2023-07-31)
 
 ### Features